move all worker id reference to ints, not longs

+++ SocialCareCaseViewerApi/V1/Boundary/Requests/UpdateMashReferral.cs
@@ -7,7 +7,7 @@ namespace SocialCareCaseViewerApi.V1.Boundary.Requests
     public class UpdateMashReferral
     {
         public string? WorkerEmail { get; set; }
-        public long? WorkerId { get; set; }
+        public int? WorkerId { get; set; }
         public string UpdateType { get; set; } = null!;
         public string? Decision { get; set; }
         public string? ReferralCategory { get; set; }
